See the commit message for the stuff I added.

### Pagination
I am not sure if we even want to do this, but I guess this makes sense if you scale beyond a handful of users, so I tried to implement something pretty basic in the endpoint for getting a list of posts. Basically this uses cursor navigation, which seems to be preferred over page navigation by most APIs.

For that, I created a Pydantic schema `Pagination` which has `after` and `before` fields, that basically act as links to the `next` and `previous` page. However, I am using timestamps for these fields, which only works for one type of sorting: descending timestamps. The official Reddit API (and most of the other APIs I have looked at) uses the `id` for these fields, which is definitely smarter, because it would allow arbitrary sorting. But I haven't quite figured out how to implement the CRUD logic around that, so this is definitely still WIP.

### Serializing neomodel
I am not sure if we really want to do it this way, but I ran into the following problem: When I want to convert the `Post` model to the `Post` schema, the relationships are not (and should not) be converted. Since the author of a post is a relationship, the `Post` schema does not receive the author without explicitly setting it. Using the `serialize` property prevents this manual process. While writing this, I thought of a case where this becomes a problem, though: If I remove a user, the post still exists. But now it doesn't have an author anymore and the `serialize` will crash :confused: So again, this is not ideal and still WIP.

### Testing
I wrote some fixtures for getting random schemas and for creating random posts and users in the database. Imho, this allows us to write tests quicker and provide a more controlled cleanup than to clear the entire database. During cleanup, the fixtures just remove what they previously added and leave the rest of the database untouched.

I much prefer this approach, because we can now technically test on the production database (still a bad idea of course, but since the community edition of Neo4j only allows one database, I have to use it for everything :D).

There is one small caveat though: cleaning up after test methods that actually test the creation of something, is not very intuitive. I found a pretty useful design pattern for this that provides fixtures that actually return a function, which can then be used by the test function to remove the created item. What bothers me: Calling this function should occur before any `assert`, because if the `assert` fails, the function is never executed. This doesn't seem to be as robust as the classic fixture system, where the cleanup is always executed.

I'd be happy about second (or more) opinions on these three topics. Feel free to contact me on Teams as well.

Related work items: #69, #71, #72

+++ backend/app/crud/base.py
@@ -1,8 +1,10 @@
+from datetime import datetime, timezone
 from typing import Any, Dict, Generic, Optional, Type, TypeVar, Union
+from uuid import UUID
 
 from fastapi.encoders import jsonable_encoder
-from neomodel import NodeSet, StructuredNode, UniqueIdProperty, db
-from pydantic import BaseModel
+from neomodel import NodeSet, StructuredNode, db
+from pydantic import UUID4, BaseModel
 
 ModelType = TypeVar("ModelType", bound=StructuredNode)
 CreateSchemaType = TypeVar("CreateSchemaType", bound=BaseModel)
@@ -20,13 +22,16 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         self.model = model
 
     @db.read_transaction
-    def get(self, uid: UniqueIdProperty) -> Optional[ModelType]:
+    def get(self, uid: Union[str, UUID4]) -> Optional[ModelType]:
         """
         Get one node from the model class.
         :param uid: the unique identifier of the node to get
         :return: the requested node if it exists, else None
         """
-        return self.model.nodes.get_or_none(uid=uid)
+        if isinstance(uid, UUID):
+            return self.model.nodes.get_or_none(uid=uid.hex)
+        else:
+            return self.model.nodes.get_or_none(uid=uid)
 
     @db.read_transaction
     def get_all(self) -> NodeSet:
@@ -44,7 +49,7 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         :return: the database model of the created node
         """
         obj_in_data = jsonable_encoder(obj_in)
-        db_obj = self.model(**obj_in_data)  # type: ignore
+        db_obj = self.model(**obj_in_data)
         db_obj.save()
         return db_obj
 
@@ -58,24 +63,30 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         :param obj_in: the UpdateSchema of the node to be updated
         :return: the updated database model
         """
-        obj_data = jsonable_encoder(db_obj)
         if isinstance(obj_in, dict):
             update_data = obj_in
         else:
             update_data = obj_in.dict(exclude_unset=True)
-        for field in obj_data:
+        for field in db_obj.__properties__:
             if field in update_data:
                 setattr(db_obj, field, update_data[field])
+        if hasattr(db_obj, "updated_at"):
+            setattr(db_obj, "updated_at", datetime.now(timezone.utc))
         db_obj.save()
         return db_obj
 
     @db.write_transaction
-    def remove(self, uid: UniqueIdProperty) -> Optional[ModelType]:
+    def remove(self, uid: Union[str, UUID4]) -> Optional[ModelType]:
         """
         Remove a node.
         :param uid: the unique identifier of the node to be removed
         :return: the removed node
         """
-        obj = self.model.nodes.get_or_none(uid=uid)
-        obj.delete()
+        if isinstance(uid, UUID):
+            obj = self.model.nodes.get_or_none(uid=uid.hex)
+        else:
+            obj = self.model.nodes.get_or_none(uid=uid)
+
+        if obj is not None:
+            obj.delete()
         return obj
